+++ app/screens/Index.js
@@ -1,4 +1,3 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import React, { useState } from 'react';
@@ -29,25 +28,25 @@ export default () => {
     //     saveToDeviceBalanceData();
     // }, [totalbalance])
     
-    const saveToDeviceBalanceData = async (totalbalance) => {
-        try {
-            const stringyfyBalanceData = JSON.stringify(totalbalance)
-            await AsyncStorage.setItem('BalanceData', stringyfyBalanceData)
-        } catch (error) {
-            console.log(error)
-        }
-    }
+    // const saveToDeviceBalanceData = async (totalbalance) => {
+    //     try {
+    //         const stringyfyBalanceData = JSON.stringify(totalbalance)
+    //         await AsyncStorage.setItem('BalanceData', stringyfyBalanceData)
+    //     } catch (error) {
+    //         console.log(error)
+    //     }
+    // }
 
-    const GetSavedBalanceDataFromDevie = async () => {
-        try {
-            const balance = await AsyncStorage.getItem('BalanceData');
-            if (balance != null) {
-                setTotalBalance(JSON.parse(balance))
-            }
-        } catch (error) {
-            console.log(error)
-        }
-    }
+    // const GetSavedBalanceDataFromDevie = async () => {
+    //     try {
+    //         const balance = await AsyncStorage.getItem('BalanceData');
+    //         if (balance != null) {
+    //             setTotalBalance(JSON.parse(balance))
+    //         }
+    //     } catch (error) {
+    //         console.log(error)
+    //     }
+    // }
 
     return (
         <UserContext.Provider